Ingredient model cleanup and NamedEntityService
