--- src/main/java/com/jaykorhonen/foodtracker/model/IngredientBase.java
@@ -1,32 +1,0 @@
-package com.jaykorhonen.foodtracker.model;
-
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-
-@EqualsAndHashCode(callSuper = true)
-@Data
-@SuperBuilder
-@NoArgsConstructor
-abstract class IngredientBase extends NamedEntity {
-
-    double fat;
-    double carbs;
-    double protein;
-    double servingSize;
-
-    static double CalculateCalories(IngredientBase ingredient, double weight) {
-        if (ingredient == null) {
-            return 0;
-        }
-
-        double servingSize = ingredient.getServingSize();
-        double servings = weight / servingSize;
-        return servings * ingredient.getCalories();
-    }
-
-    public double getCalories() {
-        return (9*fat) + 4*(carbs + protein);
-    }
-}
