+++ src/main/java/com/jaykorhonen/foodtracker/service/ServiceBase.java
@@ -1,12 +1,13 @@
 package com.jaykorhonen.foodtracker.service;
 
-import com.jaykorhonen.foodtracker.dto.NamedDTO;
+import com.jaykorhonen.foodtracker.dto.DTOBase;
 import com.jaykorhonen.foodtracker.exceptions.IngredientAlreadyExistsException;
+import com.jaykorhonen.foodtracker.model.EntityBase;
 
 import javax.naming.directory.InvalidAttributeIdentifierException;
 import java.util.List;
 
-public interface ServiceBase<T extends NamedDTO> {
+public interface ServiceBase<T extends DTOBase, E extends EntityBase> {
 
     T create(T dto) throws IngredientAlreadyExistsException, InvalidAttributeIdentifierException;
